[ClacoForm] Cascade list (#2248)
* Adds an option to hide field

* Adds cascade select field

* Allows to create sub-list in field creation form

* Fixes a maximum level for cascade select

* Allows to edit cascade select field

* checkstyle

* Allows to select a list for sub-list

* Fixes pdf export for cascade select

* Fixes category edition checkbox

* Handles categories for answers of cascade lists

* Allows to configure categories to display in widget

* Allows to configure default fields that will appear in entries list

* Fixes widget configuration

* Adds a button to open an entry in entries list if title is not displayed

+++ plugin/claco-form/Resources/modules/ClacoForm/Controller/EntryViewCtrl.js
@@ -107,29 +107,32 @@ export default class EntryViewCtrl {
     if (this.template) {
       this.template = this.template.replace('%clacoform_entry_title%', this.entry['title'])
       this.fields.forEach(f => {
-        const name = this.ClacoFormService.removeAccent(this.ClacoFormService.removeQuote(f['name']))
-        const id = f['id']
-        let replacedField = ''
+        if (!f['hidden']) {
+          const name = this.ClacoFormService.removeAccent(this.ClacoFormService.removeQuote(f['name']))
+          const id = f['id']
+          let replacedField = ''
 
-        if (this.metadataAllowed || !f['isMetadata'] || this.isEntryOwner() || this.isShared()) {
-          switch (f['type']) {
-            case 3 :
-              replacedField = this.$filter('date')(this.entry[id], 'dd/MM/yyyy')
-              break
-            case 6 :
-              replacedField = this.entry[id] ? this.entry[id].join(', ') : ''
-              break
-            case 7 :
-              replacedField = this.FieldService.getCountryNameFromCode(this.entry[id])
-              break
-            default :
-              replacedField = this.entry[id]
+          if (this.metadataAllowed || !f['isMetadata'] || this.isEntryOwner() || this.isShared()) {
+            switch (f['fieldFacet']['type']) {
+              case 3 :
+                replacedField = this.$filter('date')(this.entry[id], 'dd/MM/yyyy')
+                break
+              case 6 :
+              case 10 :
+                replacedField = this.entry[id] ? this.entry[id].join(', ') : ''
+                break
+              case 7 :
+                replacedField = this.FieldService.getCountryNameFromCode(this.entry[id])
+                break
+              default :
+                replacedField = this.entry[id]
+            }
           }
+          if (replacedField === undefined) {
+            replacedField = ''
+          }
+          this.template = this.template.replace(`%${name}%`, replacedField)
         }
-        if (replacedField === undefined) {
-          replacedField = ''
-        }
-        this.template = this.template.replace(`%${name}%`, replacedField)
       })
     }
   }
